resume_agent/requirements.txt
streamlit
python-pptx
reportlab
python-docx
PyPDF2
python-dotenv
openai